+++ litespeed/7.4/root/51-ls-env.sh
@@ -2,7 +2,7 @@
 
 echo >&2 "Setting CS WP Env..."
 
-if [ -z ${CS_AUTH_KEY} ]; then
+if [ -z "$CS_AUTH_KEY" ]; then
   echo >&2 "CS_AUTH_KEY not set, generating random key..."
   export CS_AUTH_KEY=$(xxd -l24 -ps /dev/urandom | xxd -r -ps | base64 | tr -d = | tr + - | tr / _)
 fi