+++ litespeed/7.4/root/90-postfix.sh
@@ -5,11 +5,16 @@
 #  * SMTP_PASSWORD
 #  * SMTP_SERVER
 #  * SMTP_PORT
+#  * SMTP_MAX_SEND -- max send emails per timeframe (default 1000)
+#  * SMTP_MAX_TIMEFRAME -- max send emails timeframe (default 60s)
 #  * SMTP_FROM -- allow users to override the default from email. Can be overridden by the php mailer.
 #  * SMTP_DEFAULT_FROM_FQDN -- Image admins can set a base domain (i.e. example.com) for crafting a valid from URL.
 #                              If you choose 'example.com', then the from address will be: noreply@{container-HOSTNAME}.example.com
 #                              This is overridden by a user setting `SMTP_FROM`, or by the php mailer setting the from address.
-#  * PM_STREAM 
+#  * PM_STREAM
+
+MAX_SEND="${SMTP_MAX_SEND:-1000}"
+MAX_TIMEFRAME="${SMTP_MAX_TIME:-60s}"
 
 postfix_config() {
     cat<<EOF
@@ -38,9 +43,12 @@ inet_protocols = ipv4
 smtp_always_send_ehlo = yes
 smtp_helo_name = ${HOSTNAME}
 
-smtp_sasl_auth_enable = yes 
+smtpd_client_message_rate_limit = ${MAX_SEND}
+anvil_rate_time_unit = ${MAX_TIMEFRAME}
+
+smtp_sasl_auth_enable = yes
 smtp_sasl_password_maps = static:${SMTP_USERNAME}:${SMTP_PASSWORD}
-smtp_sasl_security_options = noanonymous 
+smtp_sasl_security_options = noanonymous
 smtp_sasl_tls_security_options = noanonymous
 smtp_tls_security_level = encrypt
 smtp_tls_loglevel = 1
@@ -58,7 +66,7 @@ EOF
 }
 
 default_from() {
-  if [ -z ${SMTP_FROM} ]; then
+  if [ -z "$SMTP_FROM" ]; then
     cat<<EOF
 www-data@${HOSTNAME} noreply@${HOSTNAME}.${SMTP_DEFAULT_FROM_FQDN}
 EOF
@@ -69,29 +77,29 @@ EOF
   fi
 }
 
-config_postfix() {  
+config_postfix() {
   postfix_config > /etc/postfix/main.cf
-  if [ -z ${SMTP_DEFAULT_FROM_FQDN} ] && [ -z ${SMTP_FROM} ]; then
+  if [ -z "$SMTP_DEFAULT_FROM_FQDN" ] && [ -z "$SMTP_FROM" ]; then
     echo "No SMTP_DEFAULT_FROM_FQDN or SMTP_FROM, skipping alias..."
   else
     default_from > /etc/postfix/default_from
     postmap /etc/postfix/default_from
     echo "smtp_generic_maps = hash:/etc/postfix/default_from" >> /etc/postfix/main.cf
-  fi  
-  if [ -z ${PM_STREAM} ]; then
+  fi
+  if [ -z "$PM_STREAM" ]; then
     echo "Skipping Postmark Message Stream configuration..."
   else
     postmark_header > /etc/postfix/postmark_header.pcre
     echo "smtp_header_checks = pcre:/etc/postfix/postmark_header.pcre" >> /etc/postfix/main.cf
   fi
   FILES="etc/localtime etc/services etc/resolv.conf etc/hosts etc/nsswitch.conf"
-  echo $HOSTNAME > /etc/mailname
+  echo "$HOSTNAME" > /etc/mailname
   for file in $FILES; do
     if [ -f /var/spool/postfix/${file} ]; then
       rm /var/spool/postfix/${file}
     fi
-    cp /${file} /var/spool/postfix/${file}
-    chmod a+rX /var/spool/postfix/${file}
+    cp "/${file}" "/var/spool/postfix/${file}"
+    chmod a+rX "/var/spool/postfix/${file}"
   done
 }